5-Archive/微信机器人开发/2026-03-05/WeChatBot-Final/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pyautogui>=0.9.54
psutil>=5.9.0
pywin32>=305
mss>=9.0.0
Pillow>=10.0.0
